Refactored console into terminal.py
