+++ flexget/plugins/cli/doc.py
@@ -6,7 +6,7 @@ import sys
 
 from flexget import options
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 from flexget.plugin import plugins
 
 log = logging.getLogger('doc')
