+++ flexget/plugins/cli/entry_list.py
@@ -7,9 +7,8 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from flexget import options
 from flexget.event import event
-from flexget.logger import console
 from flexget.manager import Session
-from flexget.terminal import TerminalTable, CLITableError, table_parser
+from flexget.terminal import TerminalTable, CLITableError, table_parser, console
 from flexget.plugins.list.entry_list import get_entry_lists, get_list_by_exact_name, get_entries_by_list_id, \
     get_entry_by_id, get_entry_by_title, EntryListList, EntryListEntry
