+++ flexget/plugins/cli/inject.py
@@ -10,7 +10,7 @@ import yaml
 from flexget import options
 from flexget.entry import Entry
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 from flexget.utils import requests
