+++ flexget/plugins/cli/perf_tests.py
@@ -5,7 +5,7 @@ import logging
 
 from flexget import options
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 from flexget.manager import Session
 
 log = logging.getLogger('perftests')
