+++ flexget/plugins/cli/search.py
@@ -6,7 +6,7 @@ from datetime import datetime, timedelta
 
 from flexget import options, plugin
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 from flexget.manager import Session
 
 def do_cli(manager, options):
