+++ flexget/plugins/cli/try_regexp.py
@@ -6,7 +6,7 @@ import logging
 
 from flexget import options, plugin
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 
 log = logging.getLogger('try_regexp')
