+++ flexget/plugins/cli/win32_service.py
@@ -10,7 +10,7 @@ import sys
 import flexget
 from flexget import options
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 
 log = logging.getLogger('win32_service')
