+++ flexget/plugins/output/memusage.py
@@ -4,7 +4,7 @@ import logging
 
 from flexget import options, plugin
 from flexget.event import event
-from flexget.logger import console
+from flexget.terminal import console
 
 try:
     from guppy import hpy
